grid layout moved outside and used as a config
