+++ src/screen.js
@@ -4,48 +4,12 @@ const path = require('path')
 const blessed = require('blessed')
 const contrib = require('blessed-contrib')
 const assetsLoader = require(path.join(__dirname, '/assetsLoader'))
-
+const { CONTAINERS_GRID_LAYOUT, SERVICES_GRID_LAYOUT, IMAGES_GRID_LAYOUT } = require('./grid.config')
 const MODES = require('../lib/modes')
 
 // @TODO should be moved outside and used as a config
 // from the user (index.js)
-const CONTAINERS_GRID_LAYOUT = {
-  'actionsMenu': [4, 4, 4, 4],
-  'help': [4, 4, 4, 4],
-  'containerInfo': [2, 2, 8, 8],
-  'containerSortList': [4, 5, 3, 2],
-  'containerList': [0, 0, 6, 10],
-  'actionStatus': [6, 0, 1, 10],
-  'containerStatus': [0, 10, 2, 2],
-  'containerUtilization': [2, 10, 3, 2],
-  'containerVsImages': [5, 10, 2, 2],
-  'containerLogs': [7, 0, 4, 12],
-  'toolbar': [11, 0, 1, 12],
-  'searchInput': [11, 0, 1, 12]
-}
 
-const SERVICES_GRID_LAYOUT = {
-  'actionsMenu': [4, 4, 4, 4],
-  'actionStatus': [6, 0, 1, 10],
-  'searchInput': [11, 0, 1, 12],
-  'help': [4, 4, 4, 4],
-  'servicesInfo': [2, 2, 8, 8],
-  'servicesList': [0, 0, 6, 10],
-  'servicesLogs': [7, 0, 4, 12],
-  'servicesStatus': [0, 10, 2, 2],
-  'servicesVsImages': [2, 10, 2, 2],
-  'toolbar': [11, 0, 1, 12]
-}
-
-const IMAGES_GRID_LAYOUT = {
-  'imageInfo': [2, 2, 8, 8],
-  'imageList': [0, 0, 6, 10],
-  'searchInput': [11, 0, 1, 12],
-  'actionStatus': [6, 0, 1, 10],
-  'help': [4, 4, 4, 4],
-  'toolbar': [11, 0, 1, 12],
-  'imageUtilization': [0, 10, 2, 2]
-}
 
 const GRID_LAYOUT = {}
 GRID_LAYOUT[MODES.container] = CONTAINERS_GRID_LAYOUT
@@ -53,7 +17,7 @@ GRID_LAYOUT[MODES.service] = SERVICES_GRID_LAYOUT
 GRID_LAYOUT[MODES.image] = IMAGES_GRID_LAYOUT
 
 class screen {
-  constructor (utils = new Map()) {
+  constructor(utils = new Map()) {
     this.mode = MODES.container
     this.screen = undefined
     this.grid = undefined
@@ -69,7 +33,7 @@ class screen {
     this.widgets = new Map()
   }
 
-  initScreen () {
+  initScreen() {
     this.screen = blessed.screen({
       title: this.title,
       fullUnicode: true,
@@ -82,7 +46,7 @@ class screen {
     this.grid = new contrib.grid({ rows: 12, cols: 12, hideBorder: true, screen: this.screen })
   }
 
-  init () {
+  init() {
     // load all hooks and widgets
     this.assets = assetsLoader.load()
 
@@ -105,7 +69,7 @@ class screen {
     this.render()
   }
 
-  initHooks () {
+  initHooks() {
     for (let [hookName, HookObject] of this.assets.get('hooks').entries()) {
       let hook = new HookObject()
       this.hooks.set(hookName, hook)
@@ -113,14 +77,14 @@ class screen {
     }
   }
 
-  clearHooks () {
+  clearHooks() {
     for (let [hookName] of this.assets.get('hooks').entries()) {
       this.hooks.delete(hookName)
       this.widgetsRepository.delete(hookName)
     }
   }
 
-  initWidgets () {
+  initWidgets() {
     const layout = GRID_LAYOUT[this.mode]
     for (let [widgetName, WidgetObject] of this.assets.get('widgets').entries()) {
       if (layout[widgetName]) {
@@ -141,38 +105,38 @@ class screen {
     }
   }
 
-  clearWidgets () {
+  clearWidgets() {
     for (let [widgetName] of this.assets.get('widgets').entries()) {
       this.widgets.delete(widgetName)
       this.widgetsRepository.delete(widgetName)
     }
   }
 
-  setWidgetsRepo () {
+  setWidgetsRepo() {
     for (let widgetObject of this.widgetsRepository.values()) {
       widgetObject.setWidgetsRepo(this.widgetsRepository)
     }
   }
 
-  setWidgetsUtils () {
+  setWidgetsUtils() {
     for (let widgetObject of this.widgetsRepository.values()) {
       widgetObject.setUtilsRepo(this.utils)
     }
   }
 
-  setWidgetsInit () {
+  setWidgetsInit() {
     for (let widgetObject of this.widgetsRepository.values()) {
       widgetObject.init()
     }
   }
 
-  renderWidgets () {
+  renderWidgets() {
     for (let widgetObject of this.widgets.values()) {
       widgetObject.renderWidget()
     }
   }
 
-  toggleMode () {
+  toggleMode() {
     const availableModes = Object.values(MODES)
     if (!availableModes.includes(this.mode)) {
       this.mode = availableModes[0]
@@ -186,7 +150,7 @@ class screen {
     }
   }
 
-  registerEvents () {
+  registerEvents() {
     this.screen.on('element focus', (curr, old) => {
       if (old && old.border) {
         old.style.border.fg = 'default'
@@ -214,7 +178,7 @@ class screen {
     })
   }
 
-  render () {
+  render() {
     // var grid = new contrib.grid({rows: 12, cols: 12, screen: this.screen})
 
     // grid.set(row, col, rowSpan, colSpan, obj, opts)
@@ -226,7 +190,7 @@ class screen {
     this.screen.render()
   }
 
-  teardown () {
+  teardown() {
     this.screen.destroy()
   }
 }
